+++ packages/jest-selectors/src/utils.ts
@@ -1,6 +1,6 @@
 import { BoundFunctions, Queries, Screen } from '@testing-library/react';
 
-import { JestSelectors } from './types';
+import { JestSelectors, SelectorFn } from './types';
 
 /**
  * Get Jest Selectors
@@ -20,7 +20,8 @@ export const getJestSelectors =
        * @param args
        */
       const getElement = (noThrowOnNotFound = false, ...args: unknown[]) => {
-        const value = typeof selector === 'function' ? selector(...args) : selector;
+        const getValue = typeof selector === 'object' && 'selector' in selector ? selector.selector : selector;
+        const value = typeof getValue === 'function' ? getValue(...args) : getValue;
 
         if (value.startsWith('data-testid') || enforceTestIdSelectorForKeys.includes(key as keyof TSelectors)) {
           return noThrowOnNotFound ? screen.queryByTestId(value) : screen.getByTestId(value);
@@ -35,3 +36,44 @@ export const getJestSelectors =
       };
     }, {} as JestSelectors<TSelectors>);
   };
+
+/**
+ * Create Selector
+ * Function overloading for correct parameters types
+ * @param selector
+ * @param propName
+ */
+function createSelector<TSelector extends string>(
+  selector: TSelector,
+  propName?: string
+): { selector: () => string; apply: () => Record<string, string> };
+function createSelector<TSelector extends SelectorFn>(
+  selector: TSelector,
+  propName?: string
+): { selector: typeof selector; apply: (...args: Parameters<TSelector>) => Record<string, string> };
+function createSelector<TSelector extends () => string>(
+  selector: TSelector,
+  propName?: string
+): { selector: unknown; apply: (...args: unknown[]) => Record<string, string> } {
+  const selectorFn = typeof selector === 'string' ? () => selector as string : selector;
+
+  let attrName = 'aria-label';
+
+  /**
+   * Attribute name for selector apply
+   */
+  if (propName) {
+    attrName = propName;
+  } else if (selectorFn().startsWith('data-testid')) {
+    attrName = 'data-testid';
+  }
+
+  return {
+    selector: selectorFn,
+    apply: (...args: Parameters<typeof selectorFn>) => ({
+      [attrName]: selectorFn(...args),
+    }),
+  };
+}
+
+export { createSelector };